Move selection properties to the Polyhedron demo directory as it is demo specific
