+++ Polygon_mesh_processing/include/CGAL/Polygon_mesh_processing/properties.h
@@ -28,9 +28,7 @@ namespace CGAL
 
 enum vertex_num_feature_edges_t { vertex_num_feature_edges };
 enum halfedge_is_feature_t      { halfedge_is_feature };
-enum vertex_selection_t         { vertex_selection};
-enum edge_selection_t           { edge_selection};
-enum face_selection_t           { face_selection};
+
 enum vertex_time_stamp_t        { vertex_time_stamp};
 enum halfedge_time_stamp_t      { halfedge_time_stamp};
 enum face_time_stamp_t          { face_time_stamp};
