+++ Polyhedron/demo/Polyhedron/properties.h
@@ -1,0 +1,12 @@
+#ifndef CGAL_POLYHEDRON_DEMO_PROPERTIES_H
+#define CGAL_POLYHEDRON_DEMO_PROPERTIES_H
+
+namespace CGAL {
+
+enum vertex_selection_t         { vertex_selection};
+enum edge_selection_t           { edge_selection};
+enum face_selection_t           { face_selection};
+
+} // namespace CGAL
+
+#endif
